New section with Grid and Flex
Add a new section using Grid where I include some cards to practise simple  grid positioning and features.

Difficulties with grid when trying to make a section with no more than three columns when bid screen is used. Solved the problem just using max-width. Try to think simple. Waiting to see if with other examples will be enough, or need to learn more advanced grid features.

Add a simple flexbox section with background, heading and button, no difficulties.

Organised all the SCSS files to make it a bit more organized. Separate typography, footer and form to other files and tried to make the BEM system simpler.

Want to use it in the next project. Will be a brainstorm to orgnize everything to reuse actual examples.